+++ app/src/main/java/com/example/matirozen/printmaxtest/retrofit/PrintmaxTestService.java
@@ -34,7 +34,7 @@ public class PrintmaxTestService {
     public static int ancho = -1;
     public static int largo = -1;
     public static int colores = -1;
-    public static double price = 0;
+    public static float price = 0;
 
     //Database
     public static UserDatabase userDatabase;
